refactor: Piece 인터페이스를 추상 클래스로 통합, AbstractPiece 제거
- Piece를 interface에서 abstract class로 변경
- AbstractPiece의 공통 구현을 Piece로 이동 후 AbstractPiece 삭제
- 모든 기물이 AbstractPiece 대신 Piece를 직접 상속하도록 변경
- Piece 생성자를 protected로 변경
